Porchlight data update
 this should be updated to Porchlight - Men's Drop-in Shelter:
http://link-dane.co/#/detail/UWDC65615_2

With the additional changes.... The drop in shelter is for 90 days not
60 (this is a change for 2016 going forward). Under shelter- It should
be called Port St. Vincent and not St. Vincent DePaul (that is the
thrift store).

This should be Porchlight - Safe Haven:
http://link-dane.co/#/detail/UWDC65615_6

With the additional changes... Under the Welcome section it is for
adults who have a documented Axis 1 mental illness and are homeless.
The hours on the bottom should be everyday 7:30am to 8pm for drop-in
guests.

Porchlight - Hospitality House: http://link-dane.co/#/detail/UWDC65615_3